Move and refactor TestSceneGameplayCursorSizeChange into Ruleset Osu Tests
